\name{CommEcol-package}
\alias{CommEcol-package}
\alias{CommEcol}
\docType{package}
\title{Community Ecology Analyses}
\description{Community Ecology Analyses.}
\details{

\tabular{ll}{
Package: \tab CommEcol\cr
Type: \tab Package\cr
Version: \tab 1.8.1\cr
Date: \tab 2024-06-18\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
Community Ecology Analyses. This should be a complement to available packages, particularly vegan.
}
\author{
Adriano Sanches Melo\cr
Maintainer: Adriano Sanches Melo <asm.adrimelo@gmail.com>
}
\keyword{ package }
