\name{japi}
\alias{japi}
\docType{data}
\title{Macroinvertebrate morphospecies living on stones of a stream in southeast Brazil}
\description{Abundance data of 66 macroinvertebrate morphospecies (columns) living on 75 stones (rows) of a stream in southeast Brazil. Stones were around 20 cm diameter and were sampled from downstream (pedra1) to upstream (pedra75) direction. Consecutive stones are spaced 1-6 m. Details can be found in Melo & Froehlich (2001).}
\usage{data(japi)}
\format{A data frame with 75 observations (stones) on 66 variables (morphospecies).}
\references{
Melo, A.S. & C.G. Froehlich. 2001.Evaluation of methods for estimating macroinvertebrate species richness using individual stones in tropical streams. Freshwater Biology 46: 711-721.}
\examples{
data(japi)
japi
}