\name{sites4.6}
\alias{sites4.6}
\docType{data}
\title{Artificial dataset containing 4 samples [rows] and 6 species[columns].}
\description{Artificial dataset containing 4 samples [rows] and 6 species[columns].}
\usage{data(sites4.6)}
\format{
  A data frame with 4 observations on the following 6 variables.
  \describe{
    \item{\code{taxon_1}}{a numeric vector}
    \item{\code{taxon_2}}{a numeric vector}
    \item{\code{taxon_3}}{a numeric vector}
    \item{\code{taxon_4}}{a numeric vector}
    \item{\code{taxon_5}}{a numeric vector}
    \item{\code{taxon_6}}{a numeric vector}
  }
}

\examples{
data(sites4.6)
sites4.6
}

